scripts/analytics/requirements.txt
google-analytics-data==0.18.8
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.2
python-dateutil==2.9.0
